lib: fix path handling issues
Due to VBoxManage not being able to mount vmdk disks pointing to a
backing storage using absolute paths, there needs to be a workaround for
when the vmdk and img are not it the project root. This makes sure that
the img and vmdk are in the same directory, preventing problems that may
occur otherwise on Windows.

This required a change to the config spec.